backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend for linker"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.3.6",
    "Django>=5.0.1",
    "django-celery-beat>=2.6.0",
    "django-debug-toolbar>=4.3.0",
    "django-enumfields>=2.1.1",
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",
    "djangorestframework-gis>=1.0",
    "djangorestframework>=3.14.0",
    "geopy>=2.4.1",
    "gunicorn>=21.2.0",
    "ipython>=8.23.0",
    "openpyxl>=3.1.2",
    "Pillow>=10.2.0",
    "psycopg>=3.1.18",
    "redis>=5.0.1",
    "requests>=2.31.0",
    "watchdog>=3.0.0",
    "django-stubs>=5.2.7",
    "django-stubs-ext>=5.2.7",
]

[dependency-groups]
dev = [
    "celery-types>=0.23.0",
    "djangorestframework-stubs>=3.16.5",
    "mypy>=1.18.2",
    "mypy-extensions>=1.1.0",
    "ruff>=0.14.2",
    "types-python-dateutil>=2.9.0.20251008",
    "types-redis>=4.6.0.20241004",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff]
extend-exclude = ["migrations"]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E", "W", "I", "F", "SIM", "UP", "TID", "DJ"]
ignore = ["DJ001"]

[tool.uv]
package = false

[tool.mypy]
strict = true
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "linker.settings"

[[tool.mypy.overrides]]
module = [
    "*.migrations.*",
    "migrations.*",
    "settings",
    "settings.*",
]
ignore_errors = true

